backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas>=2.2.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Background Tasks (for future)
celery==5.3.4
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1